requirements.txt
ffmpeg-python
openai-whisper
torch
transformers
accelerate